Complete the jwt token authentication part

+++ components/utils/AuthLogout.tsx
@@ -1,8 +1,11 @@
-import axios from "axios";
+"use client"
 import { toast } from "react-toastify";
+import Cookies from "js-cookie";
+import { AxiosRequests } from "./axiosRequests";
 
-const handleLogout = async (router: any) => {
-  const accessToken = localStorage.getItem("accessToken") as string;
+export const HandleLogout = async (router: any) => {
+  const accessToken = localStorage.getItem("access_token");
+  const reqInstance = AxiosRequests();
   if (!accessToken) {
     router.push("/");
     return;
@@ -10,19 +13,29 @@ const handleLogout = async (router: any) => {
 
   const parsedAccessToken = JSON.parse(accessToken);
   const headers = { Authorization: `Bearer ${parsedAccessToken}` };
-
+  const refresh_token = Cookies.get("refresh_token")
+  const formData = { "refresh_token": refresh_token }
   try {
-    await axios.post("http://localhost:3000/api/users/logout", null, {
+    const res = await reqInstance.post("/accounts/logout", formData, {
       headers,
       withCredentials: true,
     });
-    localStorage.clear(); // Clear all items in localStorage
-    toast.success("Logout successful");
-    router.push("/");
-  } catch (error) {
-    console.error("Logout error:", error);
-    toast.error("An error occurred during logout");
+    if (res.status === 200) {
+      localStorage.clear();
+      Cookies.remove('refresh_token')
+      toast.success("Logout successful");
+      router.push("/");
+    }
+  } catch (error: any) {
+    if (error.response.status === 401) {
+      localStorage.removeItem("access_token");
+      localStorage.removeItem("email");
+      localStorage.removeItem("name");
+      router.push("/signin");
+      toast.error("Session expired");
+    } else {
+      console.error("Logout error:", error.response);
+      toast.error("An error occurred during logout");
+    }
   }
 };
-
-export default handleLogout;
